test(racetest): racetest 내 시도횟수 관련 테스트 RoundNumberTest로 이동
Race 내 시도 횟수 관리 로직 RoundNumber 클래스로 이동하여 racetest 내 시도 횟수 관련 테스트를 RoundNumberTest로 이동
